// src/api/generated/models/AppUserRequestDTO.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type AppUserRequestDTO = {
    name?: string;
    lastName?: string;
    patronymic?: string;
    login?: string;
    password?: string;
    phoneNumber?: string;
};